Refactor IsTestOutput into an extension method
Replaced the static IsTestOutput method in Day16 with an extension
method for the Action<string[], bool>? delegate type. Introduced
a new VisualizerExtensions class in the AdventOfCode.Solutions.Core
namespace to house the extension method. Updated Day16 to use the
new extension method for determining test output. Added the
VisualizerExtensions.cs file to encapsulate this functionality.